+++ core/dbt/task/compile.py
@@ -38,8 +38,11 @@ class CompileTask(GraphRunnableTask):
         return True
 
     def get_selection_spec(self) -> SelectionSpec:
+        default_selector = self.config.get_default_selector()
         if self.args.selector_name:
             spec = self.config.get_selector(self.args.selector_name)
+        elif not (self.args.select or self.args.exclude) and default_selector:
+            spec = default_selector
         else:
             spec = parse_difference(self.args.select, self.args.exclude)
         return spec