+++ core/dbt/task/freshness.py
@@ -141,9 +141,12 @@ class FreshnessTask(GraphRunnableTask):
         processing graph. A SelectionSpec describes what nodes to select
         when creating queue from graph of nodes.
         """
+        default_selector = self.config.get_default_selector()
         if self.args.selector_name:
             # use pre-defined selector (--selector) to create selection spec
             spec = self.config.get_selector(self.args.selector_name)
+        elif not (self.args.select or self.args.exclude) and default_selector:
+            spec = default_selector
         else:
             # use --select and --exclude args to create selection spec
             spec = parse_difference(self.args.select, self.args.exclude)