+++ core/dbt/task/list.py
@@ -172,8 +172,11 @@ class ListTask(GraphRunnableTask):
             return self.args.select
 
     def get_selection_spec(self) -> SelectionSpec:
+        default_selector = self.config.get_default_selector()
         if self.args.selector_name:
             spec = self.config.get_selector(self.args.selector_name)
+        elif not (self.args.select or self.args.exclude) and default_selector:
+            spec = default_selector
         else:
             spec = parse_difference(self.selector, self.args.exclude)
         return spec